Remove test suite
Our initial understanding with how we could test the
plugin is thought to be wrong because:

* The Clang API doesn't seem to have a link target
for CMake

* Clang plugins are dynamically loaded and linked
at run time by the compiler that it is being
tossed into

Because of those two reasons, I do not think it is
possible for us to have a test suite like this.

So, it is being removed, and we will need to find a
way to make meaningful integration tests to test this
stuff.

Note: I suspect it was linking and running properly
with the gcc suite because gcc optimized out the
linkage to the Clang API since it was not being
called at all.